moved the dev data fixes into 'create_skeleton_data' because it makes more sense there as it's all setting up the fake dev sirius db and should not be run on real data
import tweaks

remove comments
