+++ migration_steps/prepare/create_skeleton_data/app/lookups/check_lookups_in_mapping.py
